+++ src/main/java/com/demcha/compose/document/layout/BuiltInNodeDefinitions.java
@@ -11,6 +11,7 @@ import com.demcha.compose.document.layout.definitions.LineDefinition;
 import com.demcha.compose.document.layout.definitions.ListDefinition;
 import com.demcha.compose.document.layout.definitions.PageBreakDefinition;
 import com.demcha.compose.document.layout.definitions.ParagraphDefinition;
+import com.demcha.compose.document.layout.definitions.PolygonDefinition;
 import com.demcha.compose.document.layout.definitions.RowDefinition;
 import com.demcha.compose.document.layout.definitions.SectionDefinition;
 import com.demcha.compose.document.layout.definitions.ShapeContainerDefinition;
@@ -58,6 +59,7 @@ public final class BuiltInNodeDefinitions {
                 .register(new ShapeContainerDefinition())
                 .register(new TableDefinition())
                 .register(new CanvasLayerDefinition())
+                .register(new PolygonDefinition())
                 .register(new ChartDefinition());
     }
 }